+++ src/commands/hardis/org/retrieve/sources/dx.ts
@@ -1,4 +1,4 @@
-  /* jscpd:ignore-start */
+/* jscpd:ignore-start */
 import { flags, SfdxCommand } from '@salesforce/command';
 import { Messages, SfdxError } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
@@ -7,6 +7,7 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 import * as util from 'util';
 const exec = util.promisify(child.exec);
+
 import { MetadataUtils } from '../../../../../common/metadata-utils';
 import { checkSfdxPlugin } from '../../../../../common/utils';
 
@@ -18,24 +19,50 @@ Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('sfdx-hardis', 'org');
 
 export default class DxSources extends SfdxCommand {
-
   public static title = 'Retrieve sfdx sources from org';
 
   public static description = messages.getMessage('retrieveDx');
 
   public static examples = [
-  '$ sfdx hardis:org:retrieve:sources:dx',
-  '$ sfdx hardis:org:retrieve:sources:dx --sandbox'
+    '$ sfdx hardis:org:retrieve:sources:dx',
+    '$ sfdx hardis:org:retrieve:sources:dx --sandbox'
   ];
 
   protected static flagsConfig = {
-    folder: flags.string({char: 'f', default: '.',  description: messages.getMessage('folder')}),
-    tempfolder: flags.string({char: 't', default: './tmp',  description: messages.getMessage('tempFolder')}),
-    filteredmetadatas: flags.string({char: 'm', description: messages.getMessage('filteredMetadatas')}),
-    prompt: flags.boolean({char: 'z', default: true, allowNo: true,  description: messages.getMessage('prompt')}),
-    sandbox: flags.boolean({ char: 's', default: false, description: messages.getMessage('sandboxLogin')}),
-    instanceurl: flags.string({char: 'r',  description: messages.getMessage('instanceUrl')}),
-    debug: flags.boolean({char: 'd', default: false, description: messages.getMessage('debugMode')})
+    folder: flags.string({
+      char: 'f',
+      default: '.',
+      description: messages.getMessage('folder')
+    }),
+    tempfolder: flags.string({
+      char: 't',
+      default: './tmp',
+      description: messages.getMessage('tempFolder')
+    }),
+    filteredmetadatas: flags.string({
+      char: 'm',
+      description: messages.getMessage('filteredMetadatas')
+    }),
+    prompt: flags.boolean({
+      char: 'z',
+      default: true,
+      allowNo: true,
+      description: messages.getMessage('prompt')
+    }),
+    sandbox: flags.boolean({
+      char: 's',
+      default: false,
+      description: messages.getMessage('sandboxLogin')
+    }),
+    instanceurl: flags.string({
+      char: 'r',
+      description: messages.getMessage('instanceUrl')
+    }),
+    debug: flags.boolean({
+      char: 'd',
+      default: false,
+      description: messages.getMessage('debugMode')
+    })
   };
 
   // Comment this out if your command does not require an org username
@@ -52,10 +79,10 @@ export default class DxSources extends SfdxCommand {
   public async run(): Promise<AnyJson> {
     const folder = path.resolve(this.flags.folder || '.');
     const tempFolder = path.resolve(this.flags.tempfolder || './tmp');
-    const filteredMetadatas = (this.flags.filteredmetadatas) ?
-                                  this.flags.filteredmetadatas.split(',') :
-                                  MetadataUtils.listMetadatasNotManagedBySfdx();
-    const debug = this.flags.debug || false ;
+    const filteredMetadatas = this.flags.filteredmetadatas
+      ? this.flags.filteredmetadatas.split(',')
+      : MetadataUtils.listMetadatasNotManagedBySfdx();
+    const debug = this.flags.debug || false;
 
     // Check required plugins
     const powerkitRes = await checkSfdxPlugin('sfpowerkit');
@@ -73,23 +100,41 @@ export default class DxSources extends SfdxCommand {
     await fs.ensureDir(sfdxFolder);
 
     // Retrieve metadatas
-    const packageXml = path.resolve(path.join(tempFolder, 'package.xml')) ;
-    await MetadataUtils.retrieveMetadatas(packageXml, metadataFolder, true, filteredMetadatas, this, debug);
+    const packageXml = path.resolve(path.join(tempFolder, 'package.xml'));
+    await MetadataUtils.retrieveMetadatas(
+      packageXml,
+      metadataFolder,
+      true,
+      filteredMetadatas,
+      this,
+      debug
+    );
 
     // Create sfdx project
     if (fs.readdirSync(sfdxFolder).length === 0) {
       this.ux.log('[sfdx-hardis] Creating SFDX project...');
-      const createProjectRes = await exec('sfdx force:project:create --projectname "sfdx-project"');
+      const createProjectRes = await exec(
+        'sfdx force:project:create --projectname "sfdx-project"',
+        { maxBuffer: 1024 * 2000 }
+      );
       if (debug) {
         this.ux.log(createProjectRes.stdout + createProjectRes.stderr);
       }
     }
 
     // Converting metadatas to sfdx
-    this.ux.log(`[sfdx-hardis] Converting metadatas into SFDX sources in ${sfdxFolder}...`);
+    this.ux.log(
+      `[sfdx-hardis] Converting metadatas into SFDX sources in ${sfdxFolder}...`
+    );
     process.chdir(sfdxFolder);
     try {
-      const convertRes = await exec(`sfdx force:mdapi:convert --rootdir ${path.join(metadataFolder, 'unpackaged')} ${(debug) ? '--verbose' : ''}`);
+      const convertRes = await exec(
+        `sfdx force:mdapi:convert --rootdir ${path.join(
+          metadataFolder,
+          'unpackaged'
+        )} ${debug ? '--verbose' : ''}`,
+        { maxBuffer: 10000 * 10000 }
+      );
       if (debug) {
         this.ux.log(convertRes.stdout + convertRes.stderr);
       }
@@ -98,7 +143,11 @@ export default class DxSources extends SfdxCommand {
     }
 
     // Move sfdx sources in main folder
-    this.ux.log(`[sfdx-hardis] Moving temp files to main folder ${path.resolve(folder)}...`);
+    this.ux.log(
+      `[sfdx-hardis] Moving temp files to main folder ${path.resolve(
+        folder
+      )}...`
+    );
     process.chdir(prevCwd);
     await fs.copy(sfdxFolder, path.resolve(folder));
     // Remove temporary files