+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 Note: Can be used with `sfdx plugins:install hardis-group/sfdx-hardis@beta`
 
+## [0.5.1] 2021-02-04
+
+- Fixes:
+  - Add more items to metadatas not convertible to sfdx sources
+  - Issue when using --sandbox argument
+
+## [0.5.0] 2021-02-03
+
 - New command `hardis:project:audit:callincallout`: Audit sfdx project (or metadatas) sources to list all CallIns and CallOuts from Apex / Triggers code parsing
 - New command `hardis:project:audit:remotesites`: Audit sfdx project (or metadatas) sources to list all remote site settings of an org